+++ core-ng/src/main/java/core/framework/impl/web/service/PathParamParser.java
@@ -1,4 +1,4 @@
-package core.framework.impl.web.request;
+package core.framework.impl.web.service;
 
 import core.framework.json.JSON;
 import core.framework.web.exception.BadRequestException;
@@ -8,21 +8,7 @@ import static core.framework.util.Strings.format;
 /**
  * @author neo
  */
-public class PathParamParser {
-    @SuppressWarnings("unchecked")
-    static <T> T parse(String param, Class<T> valueClass) {
-        if (String.class.equals(valueClass)) {
-            return (T) param;
-        } else if (Integer.class.equals(valueClass)) {
-            return (T) toInt(param);
-        } else if (Long.class.equals(valueClass)) {
-            return (T) toLong(param);
-        } else if (valueClass.isEnum()) {
-            return (T) toEnum(param, (Class<? extends Enum<?>>) valueClass);
-        }
-        throw new Error("not supported path param type, type=" + valueClass.getCanonicalName());
-    }
-
+public final class PathParamParser {    // used by generated WebServiceController
     public static <T extends Enum<?>> T toEnum(String value, Class<T> valueClass) {
         try {
             return JSON.fromEnumValue(valueClass, value);