+++ core-ng/src/test/java/core/framework/impl/web/service/TestWebService.java
@@ -49,6 +49,10 @@ public interface TestWebService {
     @Path("/test/:id")
     void patch(@PathParam("id") Integer id, TestRequest request);
 
+    @GET
+    @Path("/test/:id/:enum")
+    TestResponse getEnum(@PathParam("id") Long id, @PathParam("enum") TestEnum enumValue);
+
     enum TestEnum {
         @Property(name = "A1")
         A,