+++ core-ng/src/test/resources/webservice-test/test-webservice-controller-get.java
@@ -6,8 +6,8 @@ public class TestWebServiceImpl$get implements core.framework.web.Controller {
     }
 
     public core.framework.web.Response execute(core.framework.web.Request request) throws Exception {
-        java.lang.Integer id = (java.lang.Integer) request.pathParam("id", java.lang.Integer.class);
-        java.util.Optional response = delegate.get(id);
+        Integer $id = core.framework.impl.web.service.PathParamParser.toInt(request.pathParam("id"));
+        java.util.Optional response = delegate.get($id);
         return core.framework.web.Response.bean(response).status(core.framework.api.http.HTTPStatus.OK);
     }